+++ docker-containers/microservice/src/register_in_service_discovery.py
@@ -28,7 +28,7 @@ def _add_arguments(parser):
                         help='Name of the subservice. It will be visible in Armada as: '
                              '[microservice_name]:[subservice_name].')
     parser.add_argument('-c', '--health_check', help="Alternative health check path for this service.", default=None)
-    parser.add_argument('--is-single-instance', action='store_true',
+    parser.add_argument('--single-active-instance', action='store_true',
                         help="Service discovery mechanisms will return max. 1 working instance of such service. "
                              "The rest will have status 'standby'.",
                         default=False)
@@ -70,7 +70,7 @@ def main():
         "service_port": service_port,
         "service_name": full_service_name,
         "service_container_port": args.port,
-        "is_single_instance": args.is_single_instance,
+        "single_active_instance": args.single_active_instance,
     }
 
     if args.health_check: